SPU: implement repeat/loop address quirks
writing to the repeat address register locks in the written value until the next key on, causing ADPCM loop start flags to be ignored. The Misadventures of Tron Bonne depends on this or it will freeze during certain cutscenes. Valkyrie Profile also apparently depends on this (for some voice acting probably?)

this does not apply if the register is written while the voice is restarting due to a key on, which updates the repeat address but does not lock it in; Re-Loaded depends on this or music/sound will not loop correctly